Add Support For Proxying Without Region
Adds a region entry of 'custom' that can be used to negate region being concatenated into the endpoint.
Moves the region-code postfix period to the validation return value.
Adds test

Most people arent going to want to proxy this using a region subdomain as it negates both how the examples shown in the wiki and how you'd expect to proxy this as part of your site.
The thinking around it seems out of kilter with reality but I've left the behaviour there to remove any issues migrating.